Add ticks to histogram
New function to Neighbour class that adds ticks on the halfway points
of an axis.

Added an optional boolean "hafway_x_ticks" to the "get_distance_histogram",
which when set to true will run the "add_halfway_ticks" function.
The default is set to False, so calling the function without argument will
not change its behavior.


Former-commit-id: 96c0f44f0c7a9c63ce95fef698f8871e5f0a7986